refactor: 클라이언트와 서버 간 id 혼선 문제로 컴포넌트 분리
